# Dispatch on traits in iterate