An attempt at using traits in `iterate` for `AbstractArray`s (Fixes https://github.com/JuliaLang/julia/issues/41945), which makes iteration for wrapper types (eg. `OffsetArray`s) use the same `iterate` method used for the parent array, consequently obtaining the same performance. Array wrapper types may opt-in to the iteration behavior of the parent through the trait `IterationStyle`, for example as

```julia
Base.IterationStyle(A::OffsetArray) = Base.IterationStyle(parent(A))
```

By default they will fall back to the `iterate` method for `AbstractArray`s, which is the present behavior.

A comparison of the performance without using the trait and with it (see #41945 for other examples):

Without defining the trait for `OffsetArray`s:
```julia
julia> r = 1:1:1000; ro = OffsetArray(r);

julia> @btime Float64[i for i in $r];
  1.621 μs (1 allocation: 7.94 KiB)

julia> @btime Float64[i for i in $ro];
  25.306 μs (1 allocation: 7.94 KiB)
```

With the trait defined:
```julia
julia> @btime Float64[i for i in $r];
  1.677 μs (1 allocation: 7.94 KiB)

julia> @btime Float64[i for i in $ro];
  1.609 μs (1 allocation: 7.94 KiB)
```

I would love some feedback on this as it is a somewhat large change, although there seems to be a significant performance benefit to this.